Move _get_items
