+++ quixstreams/state/rocksdb/windowed/transaction.py
@@ -1,4 +1,3 @@
-import itertools
 from typing import TYPE_CHECKING, Any, Iterable, Optional, cast
 
 from quixstreams.state.base.transaction import (
@@ -12,7 +11,6 @@ from quixstreams.state.rocksdb.transaction import RocksDBPartitionTransaction
 from quixstreams.state.serialization import (
     DumpsFunc,
     LoadsFunc,
-    append_integer,
     encode_integer_pair,
     int_to_bytes,
     serialize,
@@ -460,62 +458,6 @@ class WindowedRocksDBPartitionTransaction(RocksDBPartitionTransaction):
                 result.append(((start, end), self._deserialize_value(value), prefix))
 
         return result
-
-    def _get_items(
-        self,
-        start: int,
-        end: int,
-        prefix: bytes,
-        backwards: bool = False,
-        cf_name: str = "default",
-    ) -> list[tuple[bytes, bytes]]:
-        """
-        Get all items that start between `start` and `end`
-        within the specified prefix.
-
-        This function also checks the update cache for any updates not yet
-        committed to RocksDB.
-
-        :param start: Start of the range, inclusive.
-        :param end: End of the range, exclusive.
-        :param prefix: The key prefix for filtering items.
-        :param backwards: If True, returns items in reverse order.
-        :param cf_name: The RocksDB column family name.
-        :return: A sorted list of key-value pairs.
-        """
-        start = max(start, 0)
-        if start > end:
-            return []
-
-        seek_from_key = append_integer(base_bytes=prefix, integer=start)
-        seek_to_key = append_integer(base_bytes=prefix, integer=end)
-
-        # Create an iterator over the state store
-        db_items = self._partition.iter_items(
-            lower_bound=seek_from_key,
-            upper_bound=seek_to_key,
-            cf_name=cf_name,
-        )
-
-        cache = self._update_cache
-        update_cache = cache.get_updates(cf_name=cf_name).get(prefix, {})
-        delete_cache = cache.get_deletes(cf_name=cf_name)
-
-        # Get cached updates with matching keys
-        updated_items = (
-            (key, value)
-            for key, value in update_cache.items()
-            if seek_from_key < key <= seek_to_key
-        )
-
-        # Iterate over stored and cached items and merge them to a single dict
-        merged_items = {}
-        for key, value in itertools.chain(db_items, updated_items):
-            if key not in delete_cache:
-                merged_items[key] = value
-
-        # Sort and deserialize items merged from the cache and store
-        return sorted(merged_items.items(), key=lambda kv: kv[0], reverse=backwards)
 
     def _get_timestamp(self, cache: Cache, prefix: bytes) -> Optional[int]:
         if prefix in cache.values:
